[bugfix] click on starter wrong int type
forced all masks to be 8bits

+++ iss_preprocess/vis/click_on_starter.py
@@ -149,7 +149,7 @@ def load_roi(
             data = mask_img_data
             colormap=None
         viewer.add_labels(
-            data=data.astype("uint8"),
+            data=data.astype(int),
             name=mask.replace("_", " "),
             colormap=colormap,
         )
